Move EventListenerList to file
